email_analysis_without_vector/requirements.txt
streamlit==1.28.1
email-validator==2.1.0
python-dotenv==1.0.0
openai==1.3.7
pandas==2.1.3
plotly==5.17.0